requirements.txt
Django
djangorestframework
django-celery-beat
celery
redis
python-dotenv
google-generativeai
langchain-google-genai
Pillow